move elementAlignment to alignment directory and create custom hooks for handle alignemtns
